remove usage of menu. lead users directly to search page from search input
